Add AsyncSnapshot
React components can only render in a synchronous manner,
and this prevents them from easily using asynchronous code
which can make it difficult to handle some async component
cases.

As such we introduce the AsyncSnapShot that can allow us
to create a SnapShot of an Asynchronous process in order for
us to be able to render and resolve the async state by capturing
it at a moment in time for components to reference.

Add DataContext, ErrorContext, and LoadingContext

As we find ourselves consuming the AsyncSnapshot we
consider that we may not wish to inspect the entire
AsyncSnapshot just to be able to reference certain parts of
it.  For example, a loading indicator only really needs to know
when data is loading, and is not actually concerned with the
result of the actual data being loaded.

As such, we add different Contexts to represent the separate
parts of the async process for easier reference.

Add PromiseResolver, ProvideAsyncStates and AsyncSnapshotContext

The PromiseResolver is a component that can take a Promise
and automatically resolve it's lifecycle in order for those
states to be available to descendant components.

This allows us to render states of the Async process with
discrete AsyncSnapshots derived from the component.  It also
allows our Component to update reactively as soon as there
is a state change in the passed Promise.

Add PromiseBuilder component

The PromiseBuilder is a different way of utilizing the
PromiseResolver component, for a way that might feel better
to some users.  Ultimately it uses the same proven code path
of the PromiseResolver.