Add TaggedBase64 type
In order to better facilitate the data that will need to be
translated between the server and the client it is good to have
some data types that can readily represent the data we will find
online.

This commit introduces the TaggedBase64 type.  This is an
implementation of the TaggedBase64 data type in javascript so
we can ensure that our data members match and ultimately
represent the same values on the server or locally on the client.

Add TaggedBase64Text Component

We find ourselves with the need to display the Tagged Base 64
value in a Textual way.  This component is an initial attempt at
doing this.