Add RollUpEntry class
The RollUpEntry class represents the knowledge that we need in
order to curate a list of registered roll ups that we can support with an identity within the block explorer.  This currently
represents the minimum data needed for display within the
project.

Add AvatarLogo component

The Avatar Logo Component is a simple component for
displaying an img (contained within a picture element) with the
avatar className added to the picture.  In addition there are
several higher order components created for constraining the
style variants of the various avatar images.

Add RollUpSimple component

The RollUpSimple component displays an inline representation
for registered rollups containing their logo and name determined
by the given namespace.

If the rollup namespace is not registered within the curated list,
then we display that hte rollup is unregistered, with it's value.