Add Data Table Component
The DataTable component allows for the display of tabular data.
The Data Table also provides the ability to have paginated data,
and sortable columns.  However it, itself doesn't actually
perform any sorting, or pagination.  Instead it provides
mechanisms to record state so that they can be displayed
or referenced as needed.

Fixes #18