Add basic formatting components
Add Heading Components

We need to follow correct HTML standards of page layout
in order to adhere to SEO standards and requirements. One
such requirement is the need for Heading components, and
sub heading components to follow the correct order.  Every
Page needs a title, and every HTML page needs an h1
element.  So we marry these two concepts in order to
ensure that we style the page headings correctly.

Add ParagraphTextSmall Component

There are not a lot of places where a Paragraph component
is needed, but when they are present, they seem to use the
smallest text style possible.  We add this component so we
can utilize it when needed.

Add WithEdgeMargin Higher Order Component

The Style requires a responsive design that allows for a
dynamic amount insets on the outside of the page.  These
insets seem to be applied to several different element
types. As such we need to represent this, while also
allowing for it to be applied to multiple component types
while also attempting to keep our html structure simple.
Adding a Higher Order Component for Edge Margins
addresses all of these concerns.

Closes #40