Add Localized Text Rendering Components
Add bytes to NumberFormatters

The number formatter works well for rendering unit-less
numbers.  However, there are cases where we may need
units.  `bytes` just so happens to be one of these cases.

Add ByteSizeText component

We have the need to render the size of things in terms of
bytes.  This component allows us to do this with
localization concerns.  This consumes the `bytes` member
of the `NumberFormatters`.

Add functional file for Javascript Iterators

This aims at providing map, filter, reduce functionality for
Javascript Iterators, Iterables, Generators, and anything
that implements them.

Add HexText Component

We have the need of components to render hex sequences
as hex codes.  However, these sequences can be quite
long, and as such it would be best if we could truncate
them when needed.  HexText allows us to do just that.

Add FullHexText component

The FullHexTextComponent is very similar to the HexText
component.  The only difference is that it will not attempt
to truncate the hex text at all.

Add DateTimeFormatters context

The block explorer requires us to render Dates and
Durations in localized contexts and with different formats.
This DateTimeFormatters allows for us to configure these
formatters in one place so we can reference them
throughout the application.

Add DateTimeText component

the DateTimeText component allows us to display the full
date and time components of a Timestamp in a localized
manner to the user.

Add RelativeTimeText component

The design calls for the need to display the relative
duration between two timestamps.  This component should
display the major most unit to do just that.

Add TimeText component

The TimeText component only shows the time portion of a
Timestamp.  There may not be much use for this, but it is
included as the design references it.

Closes #41