Add Hex Dump Component
The design calls for a display of the data encoded as a hex
string in a box for easy consumption.  It's format is quite
odd, and not something I truly understand, but it seems to
me that the hexdump format would more-or-less meet the
requirements as outlined.

This commit adds the HexDump component to allow for the
hex data to be viewed in a scrollable format.

Closes #42