Add Labeled Button Components
The design for buttons varies their padding and styles
depending on whether the content of the button is an icon
or text.  As such, we require the need for a distinction
between the two types.  There's nothing preventing the
misuse of these components, so this is something that
**should** be required knowledge when working with
these components.  A more specific document should
come indicating the reason for these button variations.

Closes #46