Add PagePathProvider Context
In order for the StoryBook to represent the Page
experience more accurately we need the ability for the
page paths to be different versus what we expect them
to be.  As such, we require something that can provide for
us the notion of what our actual active page is.

Add PathResolverProvider Context

In addition to the PagePathContext needing to provide
some added context of what actual page type we are on
we also need something that can tell us what our
navigation link paths should be. The PathResolver will
assist in generating link paths for reference and creating
quick links.

Add NavLink component

The NavLink component is a component that will live in the
header and provide the underlying link information and
styling for links living in the NavBar and in the Nav Drawer.

Add MainNavLinks component

The MainNavLinks component is a reuseable component
for creating the list of primary navigation NavLinks.

Add NavDrawer component

NavDrawer is a component that will house the
MainNavLinks in a slide-out drawer in a mobile form factor.

Add NavBar component

NavBar is a component that will house the MainNavLinks
in a normal nav element for a larger than mobile
experience.

Closes #10
Closes #43