Add UTC Full Date Time Formatter to Date Formatters
As part of the feedback received from the Block Explorer meeting
https://www.notion.so/espressosys/Block-Explorer-Weekly-0c9475f7f1664e09b65a49a0fe21068b?pvs=4
We potentially always need the ability to represent our
timestamps as their UTC form.  In order to achieve this, this
commit adds a variation of the default timestamp for the UTC
timezone.

Add ValueLabeled Component

The design calls for a variation on our labeled values where the
value is displayed above the corresponding label.

Add RollUpInfo Component

The RollupInfo component displays reference information
concerning a rollup, including it's Website, Block Explorer, and
Namespace.

Add RollUpAvatar40 Component

The RollUpTitle requires the Rollup's logo in a size of 40px x
40px.  The RollUpAvatar40 component provides this logo as an
avatar.

Add RollUpTitle Component

The RollUpTitle component is a larger version of the RollUpSimple
component, where the avatar image comes after the name.  It
serves as the title of the individual RollUp Page.

Add Rollups and Rollup Page

This commit adds the last two demo pages for displaying
details related to the Phase 1 MVP.

Closes #22
Closes #23