Add NextJS Package for Block Explorer
In order to effectively deploy the Block Explorer we need to be able
to serve requests for the pages via some mechanism.  Despite it
essentially being an SPA navigation may look / seem cumbersome
unless we deliver the correct pages for the correct paths.  As such
we require a server to be able to deliver the correct content.

NextJS is a perfectly acceptable solution for this, if not a little overkill
as it allows us to serve these components with reduced file sizes to
the end user.  While the content itself will not be rendered on the
server side, it does ensure that when a user is looking at an individual
page, they only download what he/she needs to.

This commit adds a package for NextJS called block-explorer to
facilitate these requirements and act as a project from which we
can generate a deployable Docker image.