Add Dockerfile
A Dockerfile is needed in order to create reproducible builds
and and deployable image.

This commit creates the Dockerfile with some standard build
processes by building the library within the Docker image itself.
It uses building images in an attempt to make the image smaller
so that dev dependencies and supporting libraries are not needed
when publishing the final image