Separate data loading from consuming
In order to have better flexibility on rendering / laying out
components, it is beneficial to separate the loading of the async
data from the displaying.  This allows us the flexibility to
customize our components when needed, for custom error or
loading components in place of the components with the actual
data.