Define Gibraltar API data types
In order to  be useable against Gibraltar, we need to understand
the data that will be consumed from Gibraltar.  This means that
we will need to source our data.  While it's not ideal, for now it
would be beneficial to run against live data using the only real
data source we have: The Hotshot Query Service.  This service
currently does not support being able to retrieve data other than
a single entry at a time.  So, we implement an idealized version
of the HotShot Query service with our desired ranged endpoints
but utilizing the underlying single entry endpoints.

This can cause our pages to take quite a while to load, as this
data needs to be loaded one request at a time.  We can still
utilize this behavior in the future, but with the added benefit of
being able to load multiple at once.