Add configuration file for block-explorer
We need the ability to specify which URL to hit, or whether or
not the block-explorer should be running with fake data generated
instead of live data.  Normally we'd do this via an Environment
variable, however Next.JS does not make this particularly easy.
In addition, we'd eventually like for this SPA to be delivered
from a different process entirely.

In this case we've made it so that it loads a configuration from
a `config.json` file so we can overwrite the behavior, should we
choose to.