Add max block fetch safe guard
If a block chain never has any transactions, we will end up
pulling every block in the block chain in order to discover this.
This is not **great**, but sadly with the Gibraltar API, we don't
really have any way of retrieving the next x transactions.  Due to
this, it makes sense to "safe-guard" pulling every single block
just to find N transactions.

This commit imposes a 50 block limit.  This could, as a
consequence, missing returning the full number of desired
transactions, even if they are available, but we're really
attempting to prevent ourselves from pulling too many blocks.
This seems like a decent trade-off.