Reorg project structure
The src/types of packages/espresso-block-explorer-components holds
far too many files and code for a concept of 'types'.  The elements
contained within are heavily varied and do not all fall under a singular
category.  As such, the project would benefit with reorganization.

This splits these concepts, as well as some previous into some more
distinct top-level concepts that are correlated.