Separate padding from Card component
The padding on the Card component is convenient and is used in the
majority of use cases.  However, there are a handful of cases where this
padding is not wanted and hinders the design.  As such it is beneficial
to separate the padding from the Card itself.  However, it is important
to maintain the existing default behavior for consumers of the
component.  As such, it is desirable to provide a separate component
for the card that removes the padding instead of having it attached.

This change separates the card padding from the card with separate
classes, while also preserving the previous default padding behavior.