Add TransactionsPerSecondText
We need a component that will display a number in terms of
`transactions per second`.  There isn't really a built in formatter for this
for `Intl`.`NumberFormat`.  As such we create our own imitation of one
by attempting to combine a supported compound unit with it's unit time
and making the assumption that the unit doesn't get changed for the
compound format.