Add comments to conversion functions
The conversion functions that act as helper functions to provide a way
to translate between retrievable availability API types to desired derived
data types are not documented.  Though what they are doing may be
fairly straight forward, it helps to provide context as to why they are
necessary in the first place.

This adds comments to these functions in order to indicate what it is
they are doing, why, and why they may no longer be necessary.