Clean up CircularBuffer tests
It is a bit complex to have to return a function from the
`basicOperations` function in order to run tests on the provided buffer.
This seems unnecessary and complex, and it is far simpler to just run
the tests on the function just by calling it instead.

This commit cleans up this complexity in hopes of being more
beneficial to readability and maintainability.