Fix warning about setState invocation
We keep receiving an error when `setState` is being invoked in a
component that is not drawing or responsible for the creation of
the `state` in question.  More specifically we are receiving this
warning from `SearchDataConsumer` which is used to trigger the
`search` calls.  This may have unknown / unintended side-effects
and this warning is indicating us of exactly that.

This commit replaces the previous approach with a React.useEffect
which helps to control the life-cycle of these events so that
they do not cause future errors.