Add user story test use case for ExplorerPage
The ExplorerPage has quite a lot of potential feature sets that add to
the benefit of the page.  We should codify this behavior into a test to
ensure that it works without failing.

Specifically this user-story is centered around the user typing a query
into the search box.  This should yield results for the user, and those
results should be navigable via the keyboard.