Add guard before creating WebWorker based service
When attempting to build the page using nextjs, it attempts to prerender
components, which causes it to evaluate some of the intended browser-side
code to be evaluated in the context of NodeJS.

`CappuccinoHotShotQueryServiceAPIContext` causes this to fail due to
creating a `Worker` immediately upon invoking the constructor of the
service.  To fix this, the context must be evaluated such that Worker cannot
be invoked, unless it exists.