Add more instructions to Readme.md
The current instructions for getting started with the project are currently
insufficient. We should include instructions that help the user to get started
from nothing, and provide instructions / commands for running the tests or
performing development tasks.

This adds instructions to the main `README.md` file that is aimed at getting
a user started and going from a freshly cloned repository.