Fix code coverage reporting for circular buffer
It turns out that the v8 coverage provider fails to accurately report the
code coverage for the CircularBuffer test. The specifics for why this is
are unknown.  However, testing CircularBuffer's tests in isolation
produces the correct coverage values, while testing the entire project
does not.

Istanbul also seems to provide a more accurate coverage.  As such this
swaps to istanbul as a coverage provider.