Improve record detection to identify errant field
When decoding an object into a class, we want to make sure we make it
as easy as possible to identify the problem area.  This means indicating
what field specifically caused the error to fail the check, should any of
them fail.  This is helpful when there are many fields that can error at the
same point.

Add new error to indicate failed object field absence