Replace repeated ../ imports with aliases
The number of up directory package imports that are in the project are a
bit overwhelming.  Couple that with the fact that these folders are
organized in such a way that they are being imported with decent
frequency lends itself to the potential for import aliases within the
project in order to clean up the relative path imports.