Add assertions to all encoders
Since Typescript is more-or-less a simple type checker at the development level it cannot
guarantee that the arguments passed to each function are in fact the arguments that were
desired.  As such, it can be helpful to put in assertions to assert that inputs to functions
are what is expected.  This allows for the input type to be checked when developing, and
these assertions can be omitted when releasing the final product.