Add codecs and testing for error types
At the moment we have the ability to decode successful responses from the external
APIs, but we don't really have a away to handle well encoded failures to those APIs.  This
is necessary in order to understand the nature of a potential failure, and as such we should
really support these cases, as they can actually come up in real world scenarios.

Unfortunately, our errors aren't currently decodable from responses received.  As such,
the first step is to ensure that our Errors are decodable, or at least for the ones that are
actually transferred in an Encodable / Decodable way.

The only real issue is that we have a circular dependency on being able to decode our
classes, specifically errors, and those decoding actions themselves having the potential
to throw errors as well.

This change adds Codecs for errors, and attempts to unwind the circular dependency.

Some potential down-sides to this is that we have introduced some basic `Error` types
that are not encodable at all.  Ideally, this types of errors shouldn't make it across the
line.  In the future, they would probably be best served as assertions.