Fix `Dockerfile` build
The build process contained within the `Dockerfile` is creating a `nextjs`
image that does not work with the `WebWorker`.  This is due to the build
files changing, and the need to manually copy over the websocket file
generated in the `assets` folder of the
`espresso-block-explorer-components` workspace.  In order to make
this future proof, and because the `public` folder consists of the results
from the components project as well as it's `public` folder, it is much
simpler to just copy those files over during the build process.