Fix Histogram missing y axis labels
When a histogram is displaying values that are all less than 1, then the
resulting y axis labels only has a single value displayed.  This makes it
difficult to determine what the values on the histogram actually are.  The
reason for this is that we're exclusively picking whole numbers on the
y-axis guidelines.

To fix this issue this removes the exclusive use of whole numbers for
y-axis labels.