Fix the guidelines being drawn beneath the y-axis labels
The guidelines are being drawn beneath the y-axis labels due to them
starting at the edge of the histogram graph.  In order to allow for some
room, they should just be offset by a set amount.  Ideally this would be
informed by the size of the text of the label.  However, with SVGs this is
somewhat difficult.  As such, an easy compromise is to just use a set
offset.