Fix SVG text scaling issue
Since the histograms are drawn as SVGs with static dimensions given, this
allows for consistent representation of the elements drawn.  Unfortunately
it also means that the desired text size doesn't match the text size of the
of the reset of the page unless the histograms are the correct size.  This is
due to the dimensions of the histogram being drawn not matching the
actual dimensions of the histogram itself.

In order to fix this, we need to know the size of the histogram after it has
been through its layout flow.  As such, we update the histograms to listen
for resize events, and for just retrieving their actual size whenever it is
mounted and available.  From there we just provide the dimensions of the
histogram as necessary.