Add tooltips to histograms
It is difficult to tell the value of the bars in the histogram just based on
inspection and the y axis labels.  Additionally the design also calls for
tooltips to be present on interaction with the histogram.  This allows for
the user to inspect a value by using their device to highlight the relevant
histogram bar.

This adds the histogram plot tooltips so that the user can inspect the
relevant values. It also includes the block number in order to identify
which value it is associated with.