Replace Throughput Histogram with Blockspace Used
The Throughput histogram is believed to potentially lead to confusion
about maximum throughput that is allowable by the Espresso Blockchain.
Additionally the data contained is somewhat lacking in terms of total
value.  As such it is recommended to replace the Transactions per second
with bytes per second, and to rename the Throughput value to something
else like `Blockspace Used`.