Fix broken tests due to `ResizeObserver` not existing
In some environments the `ResizeObserver` doesn't exists.  One such
environment is the testing environment.  To ensure that we can pass tests
in such an environment, we need to double check that the
`ResizeObserver` is a valid option.

This change checks for the `ResizeObserver` before attempting to utilize
it.  It should be fine to not have a fallback, as it just governs some minor
visual crispness.