Fix bad formatting results for BytesPerSecondNumberFormat
The BytesPerSecondNumber format still has a reference to `Tx` in it,
which isn't desired.  Furthermore, it also is incorrect in its implementation
as it doesn't reference the formatted unit without the seconds in order for
it to be replaced by the unit desired.

This change fixes these issues by using another formatter for the bytes
and using a kilometer formatter for the kilometer unit replacement.