Replace hard-coded constant with commented variable
The use of the arbitrary `60` in the code is not explained and seems to
not really indicate its purpose.  To rectify this, it has been pulled out of
the component, given a constant value, and a comment describing what
it is, what it's for, and the reason it was chosen.