Fix histogram label duplication / performance
There is a bug in the drawing of the guidelines for a histogram when there
is little variance in the numbers, and the numbers are "large".  This
culminates with many labels being overdrawn on top of each other, and
a lot of iteration in javascript which causes the page to become
unresponsive and eventually crash.

The issue has to do with not accounting for the minimum value in both
the affine transform, and in the histogram guide line drawings.