Fix histogram being entirely blank when all data is the same
When all of the data is the same in a histogram, nothing ends up being
drawn for the histogram.

To address this issue, we add some checks after computing the statistics
to see if the min and max value are the same.  If they are, we change some
of the values to ensure we have some amount of range to work with.