Fix y-axis labels overlapping, and inconsistent spacing
The y-axis labels are only given a set amount of space to draw the labels
and as such if the text grows beyond those set bounds, the data will
overlap with the information drawn.

To fix this we add some checks to measure the actual space taken up by
the y-axis labels in order to prevent other elements from taking up that
space.