Add support for missing block data in histograms
When we receive the histogram data from the API, we get a list of values
that correspond to the blocks returned.  Normally this is expected to be
contiguous data, but the server may be missing blocks causing it to skip
over blocks it doesn't have.  This means that there should be a gap in that
value.

In order to account for this, we will take the data given, sort it, and ensure
that we insert blanks for data that is missing.  We also update the
histograms to support these blanks to indicate an empty slot in the graph.

Fix plot area size

The plot area and tooltip size are calculated in an inconsistent manner.
Instead of attempting to create more hacks for adjusting the plot width,
instead we calculate the correct offsets, and replace the prior logic with
the new layout logic resulting in better layout behavior which does not
result in cutting off tooltips / bars.

Add SVGPathBuilder

In an effort to make it easier to understand the logic that is occuring
when attempting to build the tooltip card outline, we move the logic
to a new component.  Additionally we create an SVGPathBuilder with
better method and variable names to help indicate what is happening
and why.