Add BaseNumeric class for number types
Since Typescript and Javascript don't have very great type safety
guarantees, we need to impose our own type safety where possible.  To
such an end, a base number class that can be used anywhere a number
can is preferred.  It just so happens that Javascript has some interesting
behavior when it comes to operating on things that should be numbers.
It calls `valueOf()` on the object in question, and if that result is a number
then it will operate on that instead.  Using this fact, we can create our
own number classes that can be used in place of actual numbers. The
only downside of this approach is that built in operators and functions
that operate on numbers will return raw numbers as their result.

This adds the BaseNumeric class as a base class to extend upon. It
contains the basic implementation of general numeric values and
implements default valueOf() behavior on them, so that they can
be treated as numbers where needed.  Additionally, it makes it so
that these can types can be stacked on top of each other for further
disambiguation.