Add GeoJSON classes
In order to be able to sample data for geo coordinates for a widget to
create a map, we need to be able to load in continent data.  This data
should be able to be sampled, and should represent the underlying data
with convenience methods that allow for sampling and querying the data
in ways that will make it conducive to type safe interaction.

The GeoJSON classes have are being added to cover these elements
as much as possible.  Using them we can generate bounding boxes for
the coordinate data.  This will allow us to do polygon "hit tests" that will
allow us to draw a grid of dots in the shape of a map.  Additionally, with
the unit separation it helps to make it very clear which types of units we
are dealing with, and for what purpose.