Add Geodesic Projections
In order to effectively work with Geodesic coordinates we have a need to
be able to translate them between different representations.  This can
come in the form of various types of transforms that may or may not be
linear.  As such, a general concept of a Projection is needed that operates
on the LatLng tuple of coordinates.  These projections may need to be
reversible so that we can ascertain where we are landing in a local
coordinate space compared to the original coordinate space.

Additionally, this provides some initial Projection space implementations
for a Mercator Projection, a conversion between degrees and radians,
a generalized coordinate space projection, and finally a chain projection
which can automatically chain multiple projections together.