Add Winding Number and Edge Functions
In order to be able to test whether a point is within a region outlined by a
polygon, we need to be able to solve general within checks, otherwise
known as "hit" tests.  We can take advantage of fill algorithms to achieve
these goals, as they have been used for their efficiency in determining
whether a region within a polygon is filled in or not by using a winding
number.

This requires the use of "Edge Functions" to determine which "side" of
a vector a point is on.  Using this we can make a determination as to
when a segment of a polygon has crossed from one side to another.
If we assume that that line is "infinitely long" (without actually being
infinitely long), then we are able to determine at which point the
vector was crossed.

In conjunction with the Edge function to determine whether a point is
within a polygon or not, we can use this to calculate the winding number
of the point using Sunday's winding number algorithm.  This works by
fixing the vector in a single direction (pointing to the right) and testing
these points accordingly.