Add World Map Grid computations
In order to display a grid of cells as a map we need an efficient way to
represent the data in an efficient manner.  Additionally we need a way
to compute this grid from a set of constants.  This computation can be
quite taxing on resources, and it isn't able to compute data in a
reasonable amount of time.

The raw continents data is around 10 MB in size.  This would increase the
bundle size by far too much.  In addition to the computation time taking
too long it is infeasible to compute / render this data as a component in
real / reactive time.

Due to this, we can pre-compute the dots in a resolution that has many
factors.  This allows the pre-computed data to be efficiently rendered,
and even resampled efficiently.

A script is also added in order to recompute / recalculate the hit points
with a dedicated script in the package.json file.