Add Channel interface
The channel interface is meant to bridge the gap between async events
being sent in a "hot" manner.  A "hot" manner indicates that they are
delivered without being prompted. The channel then allows for the
consumption of the events in a "cold" manner.  This allows for a callback
to be translated into poll based iteration.

Add BufferedChannel

The BufferedChannel is an implementation of a Channel that allows for a
buffered number of entries to be written before being consumed.