Optimize BufferedChannel synchronization
The current implementation of BufferedChannel creates and resolves a
Completer with every poll and publish.  This creates a lot of overhead
that practically eliminates any advantage from the channel being Buffered.
Instead, it would be more optimal to only achieve the desired
synchronization when needed.  It is also beneficial to separate the
synchronization between the publish and poll methods.

Additionally, unit tests are necessary in order to verify the correct
behavior of the buffered channel.  It should also be noted that unless
the async nature of publish and poll are treated in a proper async way
the execution order of publish and poll cannot be guaranteed.