Fix reversing of Latitude and Longitude order
The concepts of Latitude and Longitude were reversed in the program
representation due to a mistake.  This mistake was predicated on the
normal representations of latitude and longitude pairs when compared
with the Geo JSON representation.  As such, `(lat, lng)` pairs were
coded to represent `(x, y)` coordinates instead of `(y, x)` coordinates.

This representation error has been corrected in the code, and the
methods and functions that touch these concepts have been modified
accordingly.