Add AsyncIterableResolver component
A component that is able to listen a continuous stream of asynchronous
updates is needed.  Much like the `PromiseResolver` component, it
needs to have the flexibility to provides updates to the underlying children
via the same React Contexts.  It is also able to provide these updates via
the same `AsyncSnapshot`, allowing for components that consume
`PromiseResolver` contexts to easily swap in to this new component.