Separate the SVG tooltip for better reuse
The Histogram Tooltip works fairly well and looks good.  However, it is
defined entirely within the context of the histogram and that doesn't
lend itself to reuse very well.  The Tooltip itself should be separated from
the Histogram so it can be reused in a general way.

This separates the ToolTip logic portions from the histogram, while
maintaining the current results compatibility.  This will allow for the
ToolTip element to be reused for different graphs and plot types.