Rename web worker files for Hot Shot Query Service
The filenames for the Hot Shot Query Service Web Worker
implementations are too generic and don't reflect that they are
specifically for the service's implementations.  As more web workers
will be added in the future it will be a point of confusion if their
names are all the same.

As such, we rename these files to have a prefix in order to indicate
that they are for the Hot Shot Query Service's implementation.