Add immutableIterator and immutableIterable to CircularBuffer
There is a need to be able to maintain a CircularBuffer's data and iterate
over it without removing it from the structure.  As such we need a "read
only iterable / iterator implementation for a CircularBuffer.

Additionally, in order to better separate these concepts the Iterable and
Iterator implementations for the CircularBuffer have been created as
a separate classes with appropriate names.