Refactor bounding box calculators to avoid slice
Since the `slice` method on arrays will actually duplicate the array it is
invoked upon, it is actually quite expensive to call just to skip certain
elements.  As such it is better to approach the processing to use an
iterative method that doesn't clone the object in question unnecessarily.

The bounding box calculations has been modified to take advantage of
iterators just for this purpose.  As a consequence these function
definitions become more flexible.

This change was suggested by @ImJeremyHe in these comments:
https://github.com/EspressoSystems/espresso-block-explorer/pull/110#discussion_r1671565957
https://github.com/EspressoSystems/espresso-block-explorer/pull/110#discussion_r1671565524