Add interaction and state tests for Block Size Histogram
It is beneficial to more clearly document the states and visual differences
that the histograms can undergo.  As such, it is beneficial to have some
storybook tests that can demonstrate the different states that
the  histogram can be in.

Additionally it serves a great purpose to ensure that the interactions with
the histogram are tested correctly.  Namely that hovering over the bars
displays the tooltips correctly.

This is being added to provide better behavior tests with some of our more complicated components.  It is also motivated by feedback / questions from @alysiahuggins on this comment:
https://github.com/EspressoSystems/espresso-block-explorer/pull/110#discussion_r1672296705