Add reference to card.css for histogram_plot.css
When refreshing in storybook there are scenarios where the tooltips
display all black.  This is due to the colors for card.css being referenced
in order to reflect the card styles in the tooltip without them being directly
loaded.

To fix this we just need to ensure that card.css is always loaded whenever
the Histogram is being displayed.