Fix bug in takeIterable / takeIterator implementation
The takeIterable and takeIterator implementation have a bug that makes them
consume too many elements of hte underlying iterable / iterator.  In order to test
this the conditions for the accumulator iteration and the consumption iteration
calls have been separated.  Additionally a test case has been added to test this
case in order to prevent future regression.