Integrate with node validator server api
The Request and Response types have been modified to match what is currently
implemented in the node_metrics package on the espresso-sequencer repo.
These changes are pervasive, and they affect other aspects of the elements. As
such existing code needed to be updated in order to respect and respond to these
new types.

Additionally, two more implementations of the node validator web worker have
been added.  One that replays recorded HAR files that capture WebSocket traffic,
this is good for testing specific response orders, and to verify correct behavior.
The other implementation added is a WebSocket implementation that controls the
actual communication with the live WebSocket server.