Remove direct access to sub class enum constructor
The CappuccinoAPIBitVecOrder class had its subclasses with direct access to
the class for instanceof checks and constructor invocation.  In the case of enums
it makes more sense for these to have a singleton instance of each variant where
applicable.  As such, static getters have been added to CappuccinoAPIBitVecOrder
and the classes themselves have been removed from being public.