Update Node page layout to relocate Pie charts
The original design for the Node Validator Dashboard had a section for the number
of nodes being reported over time.  This was removed, as it would be informed by
the Stake Table, and would largely be static most of the time.  However, now there
is a gap in the components where the data would have gone.  After a discussion
with @clu8 and Lucca, we've opted to move the Pie Charts in this place instead.