Add CDN Status Component
An element that displays the CDN status currently exists in the NodesPage, however it does not
match the design specification.  Currently there is nothing that is able to provide the current
CDN Status available to the Node Validator API, and as such it's not possible to currently retrieve
the current CDN Status.

Due to this, and also because the design does not provide any other state than a positive "Online"
state, this component is set to always display "Online".

This component is also made to be response to layout changes from full screen to medium sizes
to account for better layout configuration.